requirements.txt
django
djangorestframework
python-decouple
django-filter
psycopg2-binary
drf-yasg
Pillow
python-slugify
djangorestframework-simplejwt


